Cleaned up some comments
