+++ package.json
@@ -13,11 +13,11 @@
     "@types/react-dom": "^17.0.7",
     "@types/react-router-dom": "5.3.3",
     "express": "^4.17.1",
-    "prove-mobile-auth": "file:prove-mobile-auth-2.7.2.tgz",
+    "prove-mobile-auth": "file:prove-identity-mobile-auth-3.1.0.tgz",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
     "react-router-dom": "^5.3.1",
-    "react-scripts": "4.0.3",
+    "react-scripts": "^5.0.1",
     "typescript": "^4.3.4",
     "web-vitals": "^1.1.2"
   },
